refactor(pm): split pm_evaluator.cpp — extract cron + helpers into pm_process_markets.cpp
pm_evaluator.cpp had grown to ~3580 lines and instantiates ~15 boost::multi_index
containers in one translation unit, which makes it the heaviest TU in libraries/chain
after database.cpp and pushes peak compiler memory high enough to OOM on RAM-constrained
builders (and slows every incremental rebuild + kills build parallelism).

This is PURE CODE MOTION — no behavioural change:

- New internal header include/graphene/chain/pm_evaluator_detail.hpp declares the ~24
  shared helpers (median, settle_market, liquidate_position, cascade_liquidate,
  settle_liquidity, gc_market, route_pool_lp_return, …) in namespace
  graphene::chain::pm_detail. Declared once, defined once — no ODR/duplicate symbols.

- New TU pm_process_markets.cpp holds the helper DEFINITIONS (moved verbatim out of the
  former anonymous namespace into pm_detail) plus database::process_pm_markets() (the
  ~750-line per-block cron). This is where most of the multi_index instantiations live.

- pm_evaluator.cpp keeps the operation evaluators (do_apply) only; it includes the new
  header and pulls the helpers in with `using namespace pm_detail;`, so no call sites change.

The only non-comment edit to any moved line is dropping `inline` from pm_rt_bucket (it now
has external linkage). A line-by-line multiset diff of the moved code vs the original is
otherwise identical.

Verified: both new TUs pass `c++ -fsyntax-only` cleanly against the chain build flags.
Full object build/link not run in my environment (2GB RAM cap; database.cpp also needs a
>2GB TU), but the change is header-declared/single-defined by construction.
